Include types in main file
